agents/adk1001/pyproject.toml
[project]
name = "adk1001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.1.0",
]
